requirements.txt
python-nmap
os
sys
colorama
pyfiglet
termcolor
socket